Ultralize color info outputs and according utility functions

+++ include/vivid/color.h
@@ -28,7 +28,7 @@ class Color
         Color( const uint8_t index );
         Color( const std::string& hexOrName );
         Color( const col_t& value, const Space );
-        Color( const float r, const float g, const float b );
+        Color( const uint8_t r, const uint8_t g, const uint8_t b );
 
         bool valid() const;
         const col_t& value() const { return value_; }
